Split old reloptions logic into PG part and GPDB part to ease future merging from upstream.
Currently, we use RELOPT_KIND_HEAP for relations, including both HEAP
table and AO tables to keep consistant with upstream. And we use
RELOPT_KIND_INTERNAL for kinds of relations, like AOSEG, TOAST, and etc.
