+++ src/test/regress/expected/column_compression.out
@@ -447,7 +447,8 @@ alter table ccddl add column j int encoding(compresstype=yawn);
 ERROR:  unknown compresstype "yawn"
 alter table ccddl add column j int encoding(compresstype=zlib,
 compresslevel=100000);
-ERROR:  compresslevel=100000 is out of range (should be between 0 and 9)
+ERROR:  value 100000 out of bounds for option "compresslevel" (reloptions.c:734)
+DETAIL:  Valid values are between "0" and "9".
 alter table ccddl add column j int encoding(a=b);
 ERROR:  unrecognized parameter "a"
 drop table ccddl;
@@ -470,7 +471,8 @@ ERROR:  DEFAULT COLUMN ENCODING clause cannot override values set in WITH clause
 -- Invalid encoding clauses
 create table t1 (i int encoding (compresstype=zlib, compresslevel=19))
 with (appendonly=true, orientation=column);
-ERROR:  compresslevel=19 is out of range (should be between 0 and 9)
+ERROR:  value 19 out of bounds for option "compresslevel" (reloptions.c:734)
+DETAIL:  Valid values are between "0" and "9".
 create table t1 (i int encoding (compresstype=zlib, ahhhh=boooooo))
 with (appendonly=true, orientation=column);
 ERROR:  unrecognized parameter "ahhhh"
@@ -1840,7 +1842,8 @@ ERROR:  unrecognized parameter "a"
 alter type text set default encoding (compresstype=10);
 ERROR:  unknown compresstype "10"
 alter type text set default encoding (compresstype=zlib, compresslevel=100);
-ERROR:  compresslevel=100 is out of range (should be between 0 and 9)
+ERROR:  value 100 out of bounds for option "compresslevel" (reloptions.c:734)
+DETAIL:  Valid values are between "0" and "9".
 -- dependency check on drop type
 drop type int42 cascade;
 NOTICE:  drop cascades to 2 other objects
